feat: añade relación de muchos a muchos entre entrenamientos y usuarios

+++ database/migrations/2025_05_11_223658_create_entrenamientos_usuarios_table.php
@@ -4,27 +4,26 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
+class CreateEntrenamientosUsuariosTable extends Migration
 {
-    /**
-     * Run the migrations.
-     */
-public function up()
+    public function up()
     {
-        Schema::create('entrenamientos_usuarios', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedInteger('entrenamiento_id');  // Cambiar a unsignedInteger
-            $table->unsignedBigInteger('usuario_id');
-            $table->timestamps();
+        if (!Schema::hasTable('entrenamientos_usuarios')) {
+            Schema::create('entrenamientos_usuarios', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedInteger('entrenamiento_id');
+                $table->unsignedBigInteger('usuario_id');
+                $table->timestamps();
 
-            // Definir las claves foráneas
-            $table->foreign('entrenamiento_id')->references('id_entrenamiento')->on('entrenamientos')->onDelete('cascade');
-            $table->foreign('usuario_id')->references('id')->on('users')->onDelete('cascade');
-        });
+                // Índices para las claves foráneas
+                $table->foreign('entrenamiento_id')->references('id_entrenamiento')->on('entrenamientos')->onDelete('cascade');
+                $table->foreign('usuario_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     public function down()
     {
         Schema::dropIfExists('entrenamientos_usuarios');
     }
-};
+}
